requirements.txt
pycrypto
keyring
pkcs7
requests